requirements/local.txt
asgiref==3.5.0
astroid==2.9.3
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
Django==3.2.11
django-extensions==3.1.5
django-tz-detect==0.4.0
djongo==1.3.6
graphviz==0.19.1
idna==3.3
isort==5.10.1
joblib==1.1.0
lazy-object-proxy==1.7.1
mccabe==0.6.1
numpy==1.22.1
pandas==1.4.0
platformdirs==2.5.1
pydotplus==2.0.2
pylint==2.12.2
pymongo==3.12.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
scipy==1.8.0
six==1.16.0
sqlparse==0.2.4
threadpoolctl==3.1.0
toml==0.10.2
typing_extensions==4.1.1
tzdata==2021.5
Unipath==1.1
urllib3==1.26.8
wrapt==1.13.3
